requirements.txt
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0
scipy>=1.11.0
